New Features
As promised, plot handles for a row of image patches is introduced. Furthermore, a column of example images for the IN and OUT conditions for different radii is added on the left. Both are accomplished through a new local function, 'plotImageData'. A popup, 'hPredType', has been introduced in the Timing Panel (it fits here for now, though it has nothing to do with timing and might need to be moved in the future). This allows the user to toggle between different predictability measures [two added for now, predictability (P) and compressibility (C)], which are displayed above the image patches. These are read from precomputed spreadsheets in the 'savedData' folder. As it is common to multiple protocols, the image folder ('T' for textures in these protocols) should be kept in the same directory. Trials are now expected to be mapped to the parameter spatial frequency (f), instead of orientation (o), in line with previous image protocols from the lab. Besides these, various minor improvements for aesthetics via aspect ratio are accomplished. Two helper functions, 'getImagePatches.m' and 'getMonitorDetails.m', are borrowed from the NaturalImagesGammaProject repository, with the former undergoing some deletions.